requirements.txt
itemadapter==0.7.0
PyJWT==2.6.0
requests==2.28.1
scrapy==2.8.0
selenium==4.9.0
webdriver_manager==3.8.3
